requirements.txt
streamlitrun